Extract some shared test code
